+++ cinemadb/src/main/java/com/cinemabookingsystem/cinemadb/dto/TicketPriceDTO.java
@@ -1,14 +1,12 @@
 package com.cinemabookingsystem.cinemadb.dto;
 
-import java.math.BigDecimal;
-
 import com.cinemabookingsystem.cinemadb.model.TicketType;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class TicketPriceDTO {
     private TicketType ticketType;
-    private BigDecimal newPrice;
+    private Float newPrice;
     
     public TicketPriceDTO() {
         
@@ -20,10 +18,10 @@ public class TicketPriceDTO {
     public void setTicketType(TicketType ticketType) {
         this.ticketType = ticketType;
     }
-    public BigDecimal getNewPrice() {
+    public Float getNewPrice() {
         return newPrice;
     }
-    public void setNewPrice(BigDecimal newPrice) {
+    public void setNewPrice(Float newPrice) {
         this.newPrice = newPrice;
     }
 }
